feat: deprecated된 코드 클래스 자체를 분리
