+++ server/src/main/java/sunflower/server/api/TranslationApiController.java
@@ -5,6 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -17,11 +18,12 @@ import java.net.URI;
 
 @RequiredArgsConstructor
 @RestController
+@RequestMapping("/translations")
 public class TranslationApiController {
 
     private final TranslationService translationService;
 
-    @PostMapping("/translations")
+    @PostMapping
     public ResponseEntity<Void> registerPdf(@RequestPart("file") MultipartFile file) {
         if (file.isEmpty()) {
             throw new FileEmptyException(1, "빈 파일입니다.");
@@ -34,33 +36,15 @@ public class TranslationApiController {
                 .build();
     }
 
-    @GetMapping("/translations/{id}/status")
+    @GetMapping("/{id}/status")
     public ResponseEntity<TranslationStatusResponse> checkStatus(@PathVariable("id") Long id) {
         final TranslationStatusDto dto = translationService.status(id);
         return ResponseEntity.ok(TranslationStatusResponse.from(dto));
     }
 
-    @GetMapping("/translations/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<Void> downloadBrfFile(@PathVariable("id") Long id) {
         return null;
-    }
-
-    @Deprecated
-    @PostMapping("/translate-pdf")
-    public ResponseEntity<Void> translatePdf(@RequestPart("file") MultipartFile file) {
-        if (file.isEmpty()) {
-            throw new FileEmptyException(1, "빈 파일입니다.");
-        }
-
-        try {
-            final Long location = translationService.translate(file);
-            return ResponseEntity
-                    .created(URI.create("/translations/" + location))
-                    .build();
-        } catch (Exception e) {
-            System.out.println(e);
-            throw e;
-        }
     }
 }
